HttpRequestHeader.contentLength: long → int

+++ src/httpUtil/HttpRequestHeader.java
@@ -3,11 +3,11 @@ package httpUtil;
 public class HttpRequestHeader {
     private HttpMethod method;
     private String fileName;
-    private long contentLength;
+    private int contentLength;
 
     public HttpMethod getMethod() { return method; }
     public String getFileName() { return fileName; }
-    public long getContentLength() { return contentLength; }
+    public int getContentLength() { return contentLength; }
 
     public void setMethod(HttpMethod method) {
         this.method = method;
@@ -17,7 +17,7 @@ public class HttpRequestHeader {
         this.fileName = fileName;
     }
 
-    public void setContentLength(long contentLength) {
+    public void setContentLength(int contentLength) {
         this.contentLength = contentLength;
     }
 }
